+++ src/lib/services/block/index.ts
@@ -1,11 +1,17 @@
 import { useQuery } from "@tanstack/react-query";
 import type { UseQueryOptions } from "@tanstack/react-query";
 
-import { CELATONE_QUERY_KEYS, useBaseApiRoute } from "lib/app-provider";
+import {
+  CELATONE_QUERY_KEYS,
+  useBaseApiRoute,
+  useCelatoneApp,
+  useLcdEndpoint,
+} from "lib/app-provider";
 import type { BlocksResponse } from "lib/services/types";
 import type { BlockData } from "lib/types";
 
 import { getBlockData, getBlocks } from "./api";
+import { getLatestBlockLcd } from "./lcd";
 
 export const useBlocks = (
   limit: number,
@@ -30,6 +36,21 @@ export const useBlockData = (height: number, enabled = true) => {
       enabled,
       retry: false,
       refetchOnWindowFocus: false,
+    }
+  );
+};
+
+export const useLatestBlock = () => {
+  const { currentChainId } = useCelatoneApp();
+  const endpoint = useLcdEndpoint();
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.BLOCK_LATEST_HEIGHT, endpoint, currentChainId],
+    async () => getLatestBlockLcd(endpoint),
+    {
+      retry: false,
+      refetchOnWindowFocus: false,
+      cacheTime: 0,
     }
   );
 };